refactor(front):  move delete-account-form into forms
